Move debug related classes to a common place
