+++ modules/ballerina-core/src/main/java/org/ballerinalang/util/debugger/DebugServer.java
@@ -15,7 +15,7 @@
 * specific language governing permissions and limitations
 * under the License.
 */
-package org.ballerinalang.bre.nonblocking.debugger;
+package org.ballerinalang.util.debugger;
 
 import org.ballerinalang.util.debugger.VMDebugManager;
