+++ modules/ballerina-core/src/main/java/org/ballerinalang/util/debugger/VMDebugManager.java
@@ -21,9 +21,7 @@ package org.ballerinalang.util.debugger;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.netty.channel.Channel;
 import org.ballerinalang.bre.Context;
-import org.ballerinalang.bre.nonblocking.debugger.BreakPointInfo;
-import org.ballerinalang.bre.nonblocking.debugger.DebugClientHandler;
-import org.ballerinalang.bre.nonblocking.debugger.DebugServer;
+import org.ballerinalang.util.debugger.info.BreakPointInfo;
 import org.ballerinalang.runtime.Constants;
 import org.ballerinalang.util.codegen.LineNumberInfo;
 import org.ballerinalang.util.codegen.ProgramFile;
