+++ modules/ballerina-core/src/main/java/org/ballerinalang/util/debugger/info/VariableInfo.java
@@ -1,7 +1,7 @@
 /*
 *   Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
 */
-package org.ballerinalang.bre.nonblocking.debugger;
+package org.ballerinalang.util.debugger.info;
 
 import org.ballerinalang.model.values.BJSON;
 import org.ballerinalang.model.values.BMap;
